refactor: reorganize landing page components and update import paths
- Moved various landing page components into a dedicated `components/pages/landing` directory for improved organization and maintainability.
- Updated import paths in `page.tsx` and added new components including `CTA`, `FeaturesShowcase`, `HeroBackground`, `HeroContent`, `HeroImage`, `HeroStats`, and `StatsGrid`, enhancing clarity and consistency in the codebase.
