+++ src/components/pages/landing/features-showcase.tsx
