+++ src/components/pages/landing/hero-background.tsx
