+++ src/components/pages/landing/hero-content.tsx
