+++ src/components/pages/landing/hero-image.tsx
